Replace previous tables_demo_planning by refactoring
