+++ tables_demo_planning/src/tables_demo_planning/tables_demo_text.py
@@ -2,8 +2,8 @@ from typing import Dict, List, Optional
 from collections import defaultdict
 from geometry_msgs.msg import Point, Pose
 from unified_planning.plans import ActionInstance
-from tables_demo_planning.refactor.components import ArmPose, Item, Location, Robot
-from tables_demo_planning.refactor.tables_demo import EnvironmentRepresentation, TablesDemoDomain
+from tables_demo_planning.components import ArmPose, Item, Location, Robot
+from tables_demo_planning.tables_demo import EnvironmentRepresentation, TablesDemoDomain
 
 """This script provides text-based simulation and execution methods without ROS."""
